Agentic AI/requirements.txt
openai
streamlit
python-dotenv
pytest